Deprecating Adapters.cs and Replacing with Partners.cs
