--- com.chartboost.mediation/Runtime/Adapters.cs
@@ -1,27 +1,0 @@
-namespace Chartboost
-{
-    /// <summary>
-    /// List of Adapter identifiers to be utilized when calling <see cref="ChartboostMediation.StartWithOptions"/>
-    /// </summary>
-    public static class Adapters
-    {
-        public const string AdColony = "adcolony";
-        public const string AdMob = "admob";
-        public const string AmazonPublisherServices = "amazon_aps";
-        public const string AppLovin = "applovin";
-        public const string MetaAudienceNetwork = "facebook";
-        public const string DigitalTurbineExchange = "fyber";
-        public const string GoogleBidding = "google_googlebidding";
-        public const string InMobi = "inmobi";
-        public const string IronSource = "ironsource";
-        public const string Mintegral = "mintegral";
-        public const string Pangle = "pangle";
-        public const string Tapjoy = "tapjoy";
-        public const string Unity = "unity";
-        public const string Vungle = "vungle";
-        public const string Yahoo = "yahoo";
-        public const string MobileFuse = "mobilefuse";
-        public const string Verve = "verve";
-        public const string HyprMX = "hyprmx";
-    }
-}
